% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tblbldr.R
\name{read_tblbldr}
\alias{read_tblbldr}
\title{read_tblbldr
Import CSV from TableBuilder}
\usage{
read_tblbldr(path)
}
\arguments{
\item{path}{path to the CSV to be imported}
}
\value{
a dataframe
}
\description{
read_tblbldr
Import CSV from TableBuilder
}
